% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\docType{methods}
\name{arith}
\alias{Arith,ANY,ml.col.def-method}
\alias{Arith,ml.col.def,ANY-method}
\alias{Arith,ml.col.def,ml.col.def-method}
\alias{arith}
\title{Arithmetic Operators}
\usage{
\S4method{Arith}{ml.col.def,ml.col.def}(e1, e2)

\S4method{Arith}{ml.col.def,ANY}(e1, e2)

\S4method{Arith}{ANY,ml.col.def}(e1, e2)
}
\arguments{
\item{e1, e2}{numeric vectors or string or \link{ml.col.def} object.}
}
\description{
Arithmetic Operators
}

