% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{$,ml.data.frame-method}
\alias{$,ml.data.frame-method}
\title{Returns a \link{ml.data.frame} field as a \link{ml.col.def-class}}
\usage{
\S4method{$}{ml.data.frame}(x, name)
}
\arguments{
\item{x}{an \link{ml.data.frame} object}

\item{name}{field name}
}
\value{
\link{ml.col.def-class} object
}
\description{
Returns a \link{ml.data.frame} field as a \link{ml.col.def-class}
}

