% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{colnames,ml.data.frame-method}
\alias{colnames,ml.data.frame-method}
\title{Column Names of an ml.data.frame object}
\usage{
\S4method{colnames}{ml.data.frame}(x)
}
\arguments{
\item{x}{an ml.data.frame object}
}
\description{
Column Names of an ml.data.frame object
}

