% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\docType{methods}
\name{Compare,ml.col.def,ANY-method}
\alias{Compare,ml.col.def,ANY-method}
\title{Relational Operators}
\usage{
\S4method{Compare}{ml.col.def,ANY}(e1, e2)
}
\arguments{
\item{e1}{an ml.col.def object.}

\item{e2}{any object}
}
\description{
Relational Operators
}

