% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\name{degrees}
\alias{degrees}
\title{Degrees}
\usage{
degrees(x)
}
\arguments{
\item{x}{an \link{ml.col.def-class} object.}
}
\value{
numeric expression converted from radians to degrees.
}
\description{
Returns numeric expression converted from radians to degrees. The function is applied
when the result is returned to the client.
}

