% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{dim,ml.data.frame-method}
\alias{dim,ml.data.frame-method}
\title{Dimensions of an \link{ml.data.frame} object}
\usage{
\S4method{dim}{ml.data.frame}(x)
}
\arguments{
\item{x}{an ml.data.frame object}
}
\description{
Dimensions of an \link{ml.data.frame} object
}

