% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extensions.R
\name{eval_code}
\alias{eval_code}
\title{Search against a MarkLogic Database using eval extension}
\usage{
eval_code(con, query, language = "XQuery")
}
\arguments{
\item{con}{The connection object, created with  \code{\link{rfml_connect}}}

\item{query}{The query to evaluate, expressed using XQuery or JavaScript}

\item{language}{Language used for query, can be XQuery or JavaScript}
}
\value{
A data frame containing the result from the query If no results a NULL object is returned.
}
\description{
Search against a MarkLogic Database using eval extension
}
\examples{
\dontrun{
library(rfml)
con <- rfml_connect("localhost","8000", "myuser", "mypassword")
code <- "xquery=xquery version "1.0-ml"; for $lineItem in /saleOrder/lineItems return <order>{for $item in $lineItem/lineItem/productName return <item>{$item/text()}</item>}</order>"
df <- eval_code(con, code)
}
}

