% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{head,ml.data.frame-method}
\alias{head,ml.data.frame-method}
\title{Return the First Part of an \link{ml.data.frame}}
\usage{
\S4method{head}{ml.data.frame}(x, n = 6, ...)
}
\arguments{
\item{x}{an ml.data.frame object}

\item{n}{a single integer. The number of rows to return, default is 6}

\item{...}{not used}
}
\description{
Return the First Part of an \link{ml.data.frame}
}

