% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\name{is.ml.col.def}
\alias{is.ml.col.def}
\title{Check if an object is of type ml.col.def}
\usage{
is.ml.col.def(x)
}
\arguments{
\item{x}{The input can be of any type.}
}
\value{
True if it is a ml.col.def. False otherwise.
}
\description{
This function checks if the input is of type ml.col.def.
}

