% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{is.ml.data.frame}
\alias{is.ml.data.frame}
\title{Check if an object is of type ml.data.frame}
\usage{
is.ml.data.frame(x)
}
\arguments{
\item{x}{The input can be of any type.}
}
\value{
True if it is a ml.data.frame object. False otherwise.
}
\description{
This function checks if the input is of type \link{ml.data.frame}.
}

