% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\docType{methods}
\name{Math,ml.col.def-method}
\alias{Math,ml.col.def-method}
\title{Miscellaneous Mathematical Functions}
\usage{
\S4method{Math}{ml.col.def}(x)
}
\arguments{
\item{x}{an \link{ml.col.def-class} object.}
}
\description{
Mathematical functions that can be used on \link{ml.data.frame} fields. The function is applied
when the result is returned to the client.
Only abs, acos, asin, atan, ceiling, cos, cosh, exp, floor, log, log10 ,
tan, tanh, sqrt, sin, sinh and trunc is currently supported.
}

