% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{median,ml.col.def-method}
\alias{median,ml.col.def-method}
\title{Median}
\usage{
\S4method{median}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The median
}
\description{
Returns the median of a \link{ml.data.frame} field.
}
\examples{
\dontrun{
 locConn <- ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame(locConn, collection = "iris")
 # median
 median(mlIris$Sepal.Length)
}
}

