% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{min,ml.col.def-method}
\alias{min,ml.col.def-method}
\title{Min}
\usage{
\S4method{min}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The minimum value
}
\description{
Returns the minimum value of a sequence of a ml.data.frame field.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the median
 min(mlDf$orderLines1quantityOrdered)
}
}

