% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ml.add.index}
\alias{ml.add.index}
\title{Creates or updates a Range element index.}
\usage{
ml.add.index(x, scalarType = "string", database = "Documents", host = "",
  port = "8002", adminuser = "", password = "")
}
\arguments{
\item{x}{a ml.data.frame field that the index will be created on}

\item{scalarType}{An atomic type specification. "string" is default}

\item{database}{The name of the database to create the index in. "Documents" is default.}

\item{host}{The hostname or ipadress of the MarkLogic Manage server. Default is the same as used for ml.connect.}

\item{port}{The port number of the MarkLogic Manage server. 8002 is used default}

\item{adminuser}{The username of a user that have rights to create index. Default is the same as used for ml.connect.}

\item{password}{The password. Default is the same as used for ml.connect.}
}
\value{
The function will raise a error if something goes wrong.
}
\description{
The function creates or updates a range element index on the underlying element/property of a ml.data.frame field.
The index
}
\details{
The user that is used for the login must have administration priviligies.
}

