% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs-pkg.R
\docType{class}
\name{ml.col.def-class}
\alias{ml.col.def-class}
\title{An S4 class to represent a ml.col.def.}
\description{
An S4 class to represent a ml.col.def.
}
\section{Slots}{

\describe{
\item{\code{.expr}}{A string with expresion that define the ml.col.def}

\item{\code{.parent}}{Pointer to the \link{ml.data.frame-class} object that the field belongs to}

\item{\code{.type}}{A string with the type of field}

\item{\code{.name}}{A string with name of the field}

\item{\code{.data_type}}{A string with the data type of the field}

\item{\code{.org_name}}{A character string with the original names of field}

\item{\code{.format}}{character}

\item{\code{.xmlns}}{character}

\item{\code{.aggType}}{character}
}}

