% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.collection.R
\name{ml.collections}
\alias{ml.collections}
\title{Lists all collections in a MarkLogic Database.}
\usage{
ml.collections(conn, query = "")
}
\arguments{
\item{conn}{A \link{ml.conn-class} object created by \link{ml.connect}}

\item{query}{Limit the collections based on a query. For more information about syntax see \link{ml.data.frame}}
}
\description{
Lists all collections in a MarkLogic Database.
}
\examples{
\dontrun{
 library(rfml)
 localConn <- ml.connect()
 ml.collections(localConn)
 }
}

