% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs-pkg.R
\docType{class}
\name{ml.data.frame-class}
\alias{ml.data.frame-class}
\title{An S4 class to represent a ml.data.frame.}
\description{
An S4 class to represent a ml.data.frame.
}
\section{Slots}{

\describe{
\item{\code{.name}}{A string with the internal name for the ml.data.frame}

\item{\code{.conn}}{The \link{ml.conn-class} object that was created with ml.connect}

\item{\code{.queryArgs}}{A list with parameters used to query MarkLogic Server}

\item{\code{.start}}{A integer with the index of the first result to get}

\item{\code{.nrows}}{A integer with the number of rows in the result and that will be returned}

\item{\code{.extracted}}{A logical value indicating if we have selected a subset of fields}

\item{\code{.col.name}}{A character vector with the field names}

\item{\code{.col.data_type}}{A character vector with the data types of the fields}

\item{\code{.col.org_name}}{A character vector with the original names of fields in the source documents}

\item{\code{.col.org_xpath}}{A character vector with the xpath to the original names in the source documents}

\item{\code{.col.format}}{A character vector withthe  source document format XML/JSON}

\item{\code{.col.xmlns}}{A character vector with the namespace for the source document}

\item{\code{.col.defs}}{A list of \link{ml.col.def-class} added fields}
}}

