% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ml.lm.R
\name{ml.lm}
\alias{ml.lm}
\title{Creates a linear model}
\usage{
ml.lm(form, mlDf)
}
\description{
Returns a linear model that fits the given data set. The number of input variables should be 2,
as currently only simple linear regression model is supported. The first element of the array
should be the value of the dependent variable while the other element should be the value of
the independent variable.
}
\details{
The function eliminates all pairs for which either the first element or the second element
is empty. After the elimination, if the length of the input is less than 2, the function
returns the empty sequence. After the elimination, if the standard deviation of the
independent variable is 0, the function returns a linear model with
intercept = the mean of the dependent variable, coefficients = NaN and r-squared = NaN.
After the elimination, if the standard deviation of the dependent variable is 0,
the function returns a linear model with r-squared = NaN.
}

