% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs-pkg.R
\docType{class}
\name{MLparameter-class}
\alias{MLparameter-class}
\title{An S4 class to represent MLparameter used with \link{ml.data.frame}}
\description{
An S4 class to represent MLparameter used with \link{ml.data.frame}
}
\section{Slots}{

\describe{
\item{\code{relevanceScores}}{TRUE/FALSE. If the result attributes score, confidence and fitness should be included. Default is FALSE}

\item{\code{docUri}}{Deprecated TRUE/FALSE. If the uri of the documents in the results should be included. Default is FALSE.}

\item{\code{sourceFlat}}{TRUE/FALSE if the source document are already flat, no nesting. Works only with JSON data.}
}}

