% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{names,ml.data.frame-method}
\alias{names,ml.data.frame-method}
\title{Shows field names of a ml.data.frame object}
\usage{
\S4method{names}{ml.data.frame}(x)
}
\arguments{
\item{x}{an ml.data.frame object}
}
\description{
Shows field names of a ml.data.frame object
}

