% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\docType{methods}
\name{print,ml.col.def-method}
\alias{print,ml.col.def-method}
\title{Prints information of a ml.col.def}
\usage{
\S4method{print}{ml.col.def}(x)
}
\arguments{
\item{x}{an ml.col.def object}
}
\description{
Prints information of a ml.col.def
}

