% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{print,ml.data.frame-method}
\alias{print,ml.data.frame-method}
\title{Prints information of a \link{ml.data.frame} object}
\usage{
\S4method{print}{ml.data.frame}(x)
}
\arguments{
\item{x}{an ml.data.frame object}
}
\description{
Prints information of a \link{ml.data.frame} object
}

