% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.lm.R
\name{print.mlLm}
\alias{print.mlLm}
\title{Prints information for a simnple linear model returned by \link{ml.lm}}
\usage{
\method{print}{mlLm}(x, ...)
}
\arguments{
\item{x}{a ml.lm result}

\item{...}{not used}
}
\description{
Prints information for a simnple linear model returned by \link{ml.lm}
}

