% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query_sparql}
\alias{query_sparql}
\title{Search against a MarkLogic Database using a SPARQL query}
\usage{
query_sparql(con, sparql = "")
}
\arguments{
\item{con}{The connection object, created with  \code{\link{rfml_connect}}}

\item{sparql}{The SPARQL query string. If a empty string all triples form the database will be returned.}
}
\value{
A data frame containing the result from the search. If no results a NULL object is returned.
}
\description{
Search against a MarkLogic Database using a SPARQL query
}
\examples{
\dontrun{
library(rfml)
con <- rfml.create("localhost","8000", "myuser", "mypassword")
options(useFancyQuotes =  FALSE)
sparql <- paste("SELECT ?person WHERE { ?person <http://example.org/marklogic/predicate/livesIn>",  dQuote("London")," }", sep="")
df <- query_sparql(con, sparql)
}
}

