% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query_string}
\alias{query_string}
\title{Search against a MarkLogic Database using a string query}
\usage{
query_string(con, query = "", collection = "", directory = "",
  results = 0)
}
\arguments{
\item{con}{The connection object, created with  \code{\link{rfml_connect}}}

\item{query}{The query string. If a empty string all documents form the database will be returned.}

\item{collection}{Restricts the search to one or more collections. Is optional.}

\item{directory}{Restricts the search to a directory. Is optional.}

\item{results}{The maximum number results to return. 0 means all, which is the default value.}
}
\value{
A data frame containing the result from the search. If no results a NULL object is returned.
}
\description{
Search against a MarkLogic Database using a string query
}
\examples{
\dontrun{
library(rfml)
con <- rfml.create("localhost","8000", "myuser", "mypassword")
df <- query_string(con, "india AND sweden")
df <- query_string(con, "india AND sweden", "mycollection,myothercollection", "/inmydirectory")
}
}

