% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\name{radians}
\alias{radians}
\title{Radians}
\usage{
radians(x)
}
\arguments{
\item{x}{an \link{ml.col.def-class} object.}
}
\value{
numeric expression converted from degrees to radians.
}
\description{
Returns numeric expression converted from degrees to radians. The function is applied
when the result is returned to the client.
}

