% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfml.R
\name{rfml_connect}
\alias{rfml_connect}
\title{Creates a object of class "rfml"}
\usage{
rfml_connect(host = "localhost", port = "8000", username = "admin",
  password = "admin")
}
\arguments{
\item{host}{Hostname or ipadress of the MarkLogic http server. Default to localhost.}

\item{port}{Port number of the MarkLogic http server. 8000 is used default}

\item{username}{Username. admin is default.}

\item{password}{Password admin is default.}
}
\value{
If successful a rfml object to use in subsequent operations.
}
\description{
Creates a object of class "rfml"
}
\examples{
\dontrun{
rfml <- rfml.create("localhost","8000", "admin", "admin")
}
}

