% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{save.ml.data.frame}
\alias{save.ml.data.frame}
\title{Save the data of a ml.data.frame object data to MarkLogic.}
\usage{
save.ml.data.frame(x, name, format = "json", directory = "")
}
\arguments{
\item{x}{a ml.data.frame object.}

\item{name}{The name of the object. This will be used to name the collection that the data belongs to.}

\item{format}{The format od the documents that is created, json or XML. Default is json}

\item{directory}{The directory to save the documents, needs to start and end with a /. Default saved to /rfml/[username]/[name]/}
}
\value{
A ml.data.frame object.
}
\description{
The function will create documents based on the definition of a ml.data.frame object.
It will create a new document for each result from the serach that the ml.data.frame
is based on. Document URI is generated by MarkLogic and the directory is named based
on the string rfml,the username and the name parameter, for example /rfml/admin/iris/.
The documents will belong to a collection named "name".
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 mlIrisNew <- save.ml.data.frame(mlIris, "newiris")

}
}

