% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method-statistics.R
\name{sd.pop}
\alias{sd.pop}
\title{Standard Deviation of a population}
\usage{
sd.pop(x)
}
\arguments{
\item{x}{a ml.data.frame field.}
}
\value{
The sample standard deviation
}
\description{
Returns the sample standard deviation of a population.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the Covariance
 sd.pop(mlDf$orderLines1quantityOrdered)
}
}

