% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\docType{methods}
\name{show,ml.col.def-method}
\alias{show,ml.col.def-method}
\title{Prints information of a ml.col.def}
\usage{
\S4method{show}{ml.col.def}(object)
}
\arguments{
\item{object}{an ml.col.def object}
}
\description{
Prints information of a ml.col.def
}

