% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{show,ml.data.frame-method}
\alias{show,ml.data.frame-method}
\title{Prints information of a \link{ml.data.frame} object}
\usage{
\S4method{show}{ml.data.frame}(object)
}
\arguments{
\item{object}{an ml.data.frame object}
}
\description{
Prints information of a \link{ml.data.frame} object
}

