% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{sum,ml.col.def-method}
\alias{sum,ml.col.def-method}
\title{Sum}
\usage{
\S4method{sum}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The sum
}
\description{
Returns the sum of a ml.data.frame field.
}
\examples{
\dontrun{
 ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame(collection = "iris")
 # sum
 sum(mlIris$Sepal.Length)
}
}

