% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{summary,ml.data.frame-method}
\alias{summary}
\alias{summary,ml.data.frame-method}
\title{ml.data.frame Summaries}
\usage{
\S4method{summary}{ml.data.frame}(object, digits = max(3L, getOption("digits")
  - 3L), maxsum = 7L, ...)
}
\arguments{
\item{object}{an ml.data.frame object}

\item{digits}{integer, used for number formatting}

\item{maxsum}{not used.}

\item{...}{not used.}
}
\description{
ml.data.frame Summaries
}

