% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{as.ml.data.frame}
\alias{as.ml.data.frame}
\title{Upload a data.frame object data to MarkLogic.}
\usage{
as.ml.data.frame(x, name, format = "json", directory = "")
}
\arguments{
\item{x}{a Data Frame object.}

\item{name}{The name of the object.}

\item{format}{The format od the documents that is created, json or XML. Default is json}

\item{directory}{The directory to save the documents, needs to start and end with a /. Default saved to /rfml/[username]/[name]/}
}
\value{
A ml.data.frame object.
}
\description{
The function will upload the data within a data.frame object to MarkLogic. It will create a document
for each row. Document URI is generated by MarkLogic and the directory is named based on the string rfml,
the username and the name parameter, for example /rfml/admin/iris/.
The documents will belong to a collection named "name".
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on the iris data set
 mlIris <- as.ml.data.frame(iris, "iris")
}
}

