% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{as.ml.data.frame}
\alias{as.ml.data.frame}
\title{Will upload a data.frame object data to MarkLogic.}
\usage{
as.ml.data.frame(x, name)
}
\arguments{
\item{x}{a Data Frame object.}

\item{name}{The name of the object.}
}
\value{
A ml.data.frame object.
}
\description{
The function will upload the data within a data.frame object to MarkLogic. It will create a document
for each row. The documents will belong to a collection named rfml-username-name, where username is the
loged in user and name is the provided name in the function
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on the iris data set
 dmlIris <- as.ml.data.frame(iris, "iris")
}
}

