% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{as.ml.data.frame}
\alias{as.ml.data.frame}
\title{Upload a data.frame object data to MarkLogic server.}
\usage{
as.ml.data.frame(conn, x, name, format = "json", directory = "")
}
\arguments{
\item{conn}{A ml.conn object that has a valid connection to a MarkLogic Server}

\item{x}{a Data Frame or ml.data.frame object.}

\item{name}{The name of the object.}

\item{format}{The format od the documents that is created, json or XML. Default is json}

\item{directory}{The directory to save the documents, needs to start and end with a /. Default saved to /rfml/[username]/[name]/}
}
\value{
A ml.data.frame object.
}
\description{
The function will upload the data within a data.frame object to MarkLogic Server or create new documents
in MarkLogic Server based on a ml.data.frame object.
}
\details{
When documents are created they will have a Document URI that is bulid based on the string rfml, the rowname if a data.frame or a counter
if it is a ml.data.frame, the loged in username  and name parameter, for example /rfml/admin/iris/.
The documents will belong to a collection named "name".
}
\examples{
\dontrun{
 library(rfml)
 ml.connect()
 # create a ml.data.frame based on the iris data set
 mlIris <- as.ml.data.frame(iris, "iris")
}
}

