% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{cor,ml.col.def,ml.col.def-method}
\alias{cor,ml.col.def,ml.col.def-method}
\title{Correlation}
\usage{
\S4method{cor}{ml.col.def,ml.col.def}(x, y = NULL, use = NULL,
  method = NULL)
}
\description{
Returns the Pearson correlation coefficient of a data set. The size of the input
should be 2.
}
\details{
The function eliminates all pairs for which either the first element or the second
element is empty. After the elimination, if the length of the input is less than 2,
the function returns the empty sequence. After the elimination, if the standard
deviation of the first column or the standard deviation of the second column is 0,
the function returns the empty sequence.
}

