% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\name{cov.pop}
\alias{cov.pop}
\title{Population Covariance}
\usage{
cov.pop(x, y)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{y}{a ml.data.frame field}
}
\value{
The population covariance
}
\description{
Returns the population covariance of two variables, ml.data.frame fields.
}
\details{
The function eliminates all pairs for which either the first element or the
second element is empty. After the elimination, if the length of the input is 0,
the function returns the empty sequence.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the Covariance
 cov.pop(mlDf$orderLines1quantityOrdered, mlDf$orderLines1priceEach)
}
}

