% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{max,ml.col.def-method}
\alias{max,ml.col.def-method}
\title{Max}
\usage{
\S4method{max}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The maximum value
}
\description{
Returns the maximum value of a sequence of a ml.data.frame field.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the median
 max(mlDf$orderLines1quantityOrdered)
}
}

