% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{median,ml.col.def-method}
\alias{median,ml.col.def-method}
\title{Median}
\usage{
\S4method{median}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The median
}
\description{
Returns the median of a sequence of a ml.data.frame field.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the median
 median(mlDf$orderLines1quantityOrdered)
}
}

