% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{min,ml.col.def-method}
\alias{min,ml.col.def-method}
\title{Min}
\usage{
\S4method{min}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not currently used.}
}
\value{
The minimum value
}
\description{
Returns the minimum value of a ml.data.frame field.
}
\examples{
\dontrun{
 ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame(collection = "iris")
 # min
 min(mlIris$Sepal.Length)
}
}

