% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ml_conn.R
\name{ml.connect}
\alias{ml.connect}
\title{Creates a connection to a MarkLogic REST server.}
\usage{
ml.connect(host = "localhost", port = "8000", username = "admin",
  password = "admin")
}
\arguments{
\item{host}{Hostname or ip-adress of the MarkLogic http server. Default to localhost.}

\item{port}{Port number of the MarkLogic http server. 8000 is used default}

\item{username}{Username. admin is default.}

\item{password}{Password admin is default.}
}
\value{
Nothing if sucess, otherwise an error.
}
\description{
Creates a connection to a MarkLogic REST server.
}
\examples{
\dontrun{
library(rfml)
ml.connect("localhost","8000", "admin", "admin")
}
}

