% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{ml.data.frame}
\alias{ml.data.frame}
\title{Creates a ml.data.frame object}
\usage{
ml.data.frame(query = "", fieldFilter = "", collection = c(),
  directory = c(), relevanceScores = FALSE, docUri = FALSE)
}
\arguments{
\item{query}{The query string used to define the result.}

\item{fieldFilter}{Field level filtering. Currently only the == operator are supported. Multiple field filters are separated by ,}

\item{collection}{A list of collection URI:s to filter on.}

\item{directory}{A list of directory URI:s to filter on.}

\item{relevanceScores}{TRUE/FALSE. If the result attributes score, confidence and fitness should be included. Default is FALSE}

\item{docUri}{TRUE/FALSE. If the uri of the documents in the results should be included. Default is FALSE.}
}
\value{
A ml.data.frame object.
}
\description{
This function creates an object of ml.data.frame, it is based on a search done using
the provided query, collection and/or directory parameters.
The operations that can be applied onto this class of objects are very
similar to those of data.frame. No real data is loaded into R. The data transfered
between MarkLogic Server and R is minimized.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame("setosa")
 # using search and collection filtering
 mlIris <- ml.data.frame("setosa", collection = "iris")
 # using field filter
 mlIris <- ml.data.frame(fieldFilter = "Species == setosa")
}
}

