% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{ml.data.frame}
\alias{ml.data.frame}
\title{Create a ml.data.frame object}
\usage{
ml.data.frame(query = "", collection = c(), directory = c())
}
\arguments{
\item{query}{The query string used to define the result.}

\item{collection}{A list of collection URI:s to filter on.}

\item{directory}{A list of directory URI:s to filter on.}
}
\value{
A ml.data.frame object.
}
\description{
This function creates an object of ml.data.frame, it is based on a search done using
provided query. The operations that can be applied onto this class of objects are very
similar to those of data.frame. No real data is loaded into R. The data transfered
between MarkLogic Server and R is minimized.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 df <- ml.data.frame("försäkringskassan AND kalle")
 # using search and collection filtering
 df <- ml.data.frame("försäkringskassan AND kalle", collection=c("Analytics"))
 # using only collection filtering
 df <- ml.data.frame(collection=c("Analytics"))
}
}

