% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{ml.data.frame}
\alias{ml.data.frame}
\title{Create a ml.data.frame object pointing to a ...}
\usage{
ml.data.frame(query)
}
\arguments{
\item{password}{The password admin is default.}
}
\value{
The function will raise a error...
}
\description{
This function creates an object of ml.data.frame...
}
\examples{
\dontrun{

}
}

