% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{ml.init.database}
\alias{ml.init.database}
\title{Set up a MarkLogic database for use with rfml.}
\usage{
ml.init.database(host = "localhost", port = "8000", adminuser = "admin",
  password = "admin")
}
\arguments{
\item{host}{The hostname or ipadress of the MarkLogic http server. Default to localhost.}

\item{port}{The port number of the MarkLogic http server. 8000 is used default}

\item{adminuser}{The username of a user that have rights to install options. admin is default.}

\item{password}{The password admin is default.}
}
\value{
The function will raise a error...
}
\description{
The function installs search options and transformations/mdoules that is needed
in order to use the package. You only needed to run it once for each database.
}
\details{
The database must have a REST server and a module database.

The user that is used for the login must have the  rest-admin role,
or the following privileges:
    http://marklogic.com/xdmp/privileges/rest-admin,
    http://marklogic.com/xdmp/privileges/rest-writer,
    http://marklogic.com/xdmp/privileges/rest-reader
}
\examples{
\dontrun{
ml.init.database("localhost", "8000", "admin", "admin")
}
}

