% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.col.def.R
\name{radians}
\alias{radians}
\title{Radians}
\usage{
radians(x)
}
\arguments{
\item{x}{a ml.data.frame field.}
}
\value{
numeric expression converted from degrees to radians.
}
\description{
Returns numeric expression converted from degrees to radians.
}

