% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\name{rm.ml.data.frame}
\alias{rm.ml.data.frame}
\title{Remove the data of a ml.data.frame object in MarkLogic server database.}
\usage{
rm.ml.data.frame(x, directory = "")
}
\arguments{
\item{x}{a ml.data.frame object.}

\item{directory}{Optional. The directory where the data is stored, needs to start and end with a /.}
}
\value{
A ml.data.frame object.
}
\description{
Removes the data that whas saved to MarkLogic server database using the \link{as.ml.data.frame} function.
If using a directory parameter it that call the same value needs to be provided for this function.
The function will also delete the x object form tne R environment.
}
\examples{
\dontrun{
 rm.ml.data.frame(mlIris)
}
}
\seealso{
\code{\link{ml.data.frame}}, \code{\link{as.ml.data.frame}} for uploading data, \code{\link[rfml]{as.data.frame}} for pulling data
}
\concept{
array
}

