% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{sd,ml.col.def-method}
\alias{sd,ml.col.def-method}
\title{Standard Deviation}
\usage{
\S4method{sd}{ml.col.def}(x, na.rm = NULL)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not used currently}
}
\value{
The sample standard deviation
}
\description{
Returns the sample standard deviation of a sequence of numeric values.
}
\details{
The function returns a empty value if the number of rows of the ml.data.frame
that x belongs to is less than 2.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the Covariance
 sd(mlDf$orderLines1quantityOrdered)
}
}

