% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\name{sd.pop}
\alias{sd.pop}
\title{Standard Deviation of a population}
\usage{
sd.pop(x)
}
\arguments{
\item{x}{a ml.data.frame field.}
}
\value{
The sample standard deviation of a population.
}
\description{
Returns the sample standard deviation of a population.
}
\examples{
\dontrun{
 library(rfml)
 locConn <- ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame(locConn, collection = "iris")
 # standard deviation
 sd.pop(mlIris$Sepal.Length)
}
}

