% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ml.data.frame.R
\docType{methods}
\name{[,ml.data.frame-method}
\alias{[,ml.data.frame-method}
\title{Extract subsets of a ml.data.frame}
\usage{
\S4method{[}{ml.data.frame}(x, i, j, ..., drop = NA)
}
\arguments{
\item{x}{a ml.data.frame from which to extract element(s).}

\item{i, j}{Indices specifying elements to extract. Indices are ‘numeric’ or ‘character’ vectors or empty (missing) or ‘NULL’.}

\item{drop}{Not implemented yet.}
}
\value{
A ml.data.frame object is returned
}
\description{
Extract subset of columns and/or rows of a ml.data.frame. When extracting rows a ml.col.def
referense can be used with only the equal (==) comparsion operator or a search text.
The row filtering will be used togheter with the existing search of the ml.data.frame
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on the iris data set
 mlIris <- as.ml.data.frame(iris, "iris")
 mlIris2 <- mlIris[1:3] # select first three columns
 mlIris2 <- mlIris[,1:3] # same
 mlIris2 <- mlIris[,c("Sepal.Length","Sepal.Width","Petal.Length")] # same
 mlIris2 <- mlIris[mlIris$Species=="setosa", 1:3] # select first three columns for all rows with Spieces = setosa
 mlIris2 <- mlIris[mlIris$Species=="setosa",] # select all columns for all rows with Spieces = setosa
 mlIris2 <- mlIris["setosa",] # select all columns for all rows with "setosa" in any column
}
}

