% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\docType{methods}
\name{var,ml.col.def-method}
\alias{var,ml.col.def-method}
\title{Variance}
\usage{
\S4method{var}{ml.col.def}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not used currently}
}
\value{
The sample variance
}
\description{
Returns the sample variance of two variables, ml.data.frame fields.
}
\details{
The function returns a empty value if the number of rows of the ml.data.frame
that x belongs to is less than 2.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect("localhost", "8000", "admin", "admin")
 # create a ml.data.frame based on a search
 mlDf <- ml.data.frame("corvette NEAR/1 convertible", collection = c("Analytics"))
 # return the Covariance
 var(mlDf$orderLines1quantityOrdered)
}
}

