% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-statistics.R
\name{var.pop}
\alias{var.pop}
\title{Population variance}
\usage{
var.pop(x, na.rm = FALSE)
}
\arguments{
\item{x}{a ml.data.frame field.}

\item{na.rm}{not used currently}
}
\value{
The population variance
}
\description{
Returns the population variance of two variables, ml.data.frame fields.
}
\details{
The function returns a empty value if the number of rows of the ml.data.frame
that x belongs to is less than 2.
}
\examples{
\dontrun{
 library(rfml)
 ml.connect()
 # create a ml.data.frame based on a search
 mlIris <- ml.data.frame(collection = "iris")
 # population variance
 var.pop(mlDf$orderLines1quantityOrdered)
}
}

